avginvalidratiotimelapse

+++ rocksdb/plugin/zenfs/fs/zbd_zenfs.cc
@@ -1080,12 +1080,11 @@ void ZonedBlockDevice::AddTimeLapse(int T) {
   for(auto z : io_zones){
     uint64_t invalid_size = (z->wp_ -z->start_) - z->used_capacity_;
 
-    if(z->wp_ -z->start_){
-      ratio = (invalid_size/(z->wp_ -z->start_));
+    if(z->wp_ - z->start_){
+      ratio = (double)(invalid_size/(double)(z-)>wp_ -z->start_));
     }else{
       ratio = 0.0;
     }
-    
     ratio_sum+=ratio;
   }
   double avg_invalid_ratio = ratio_sum/(io_zones.size());
